refactor(util): 💡 move tests
